requirements.txt
# Core dependencies
memvid>=0.1.3
PyPDF2>=3.0.1
openai>=1.0.0
python-docx>=0.8.11
python-dotenv>=1.0.0
beautifulsoup4>=4.12.0
lxml>=4.9.0

# Spreadsheet support (Excel, CSV)
pandas>=2.0.0
openpyxl>=3.0.0
xlrd>=2.0.0
tabulate>=0.9.0

# RTF support
striprtf>=0.0.26

# EPUB support
ebooklib>=0.18

# PowerPoint support
python-pptx>=0.6.21
